requirements.txt
fastapi
uvicorn[standard]
PyYAML
Jinja2
qrcode
python-multipart
pillow
markdown
